\name{tnsrcomp-package}
\alias{tnsrcomp-package}
\alias{tnsrcomp}
\docType{package}
\title{
\packageTitle{tnsrcomp}
}
\description{
\packageDescription{tnsrcomp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tnsrcomp}
\packageIndices{tnsrcomp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tnsrcomp}

Maintainer: \packageMaintainer{tnsrcomp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
