\name{tnsrcomp-package}
\alias{tnsrcomp-package}
\alias{tnsrcomp}
\docType{package}
\title{
\packageTitle{tnsrcomp}
}
\description{
\packageDescription{tnsrcomp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tnsrcomp}
\packageIndices{tnsrcomp}
}
\author{
\packageAuthor{tnsrcomp}

Maintainer: \packageMaintainer{tnsrcomp}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
